@@ -1,15 +1,15 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.2.71'
-    ext.anko_version = '0.10.4'
+    ext.kotlin_version = '1.3.72'
+    ext.anko_version = '0.10.8'
 
     repositories {
         google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.1'
+        classpath 'com.android.tools.build:gradle:3.6.3'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
         // NOTE: Do not place your application dependencies here; they belong
@@ -30,7 +30,6 @@ allprojects {
 // Define versions in a single place
 ext {
     // App dependencies
-    supportLibraryVersion = '27.1.1'
 
 }
 
