@@ -5,7 +5,7 @@ import android.app.ProgressDialog
 import android.os.Build
 import android.os.Bundle
 import android.os.StrictMode
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import com.dream.freshnews.util.debug
 
 open abstract class BaseActivity : AppCompatActivity() {
